Fix Permissions#fromBitmask methods.

+++ javacord-api/src/main/java/org/javacord/api/entity/permission/Permissions.java
@@ -89,7 +89,7 @@ public interface Permissions {
      * @param bitmask The bitmask of allowed permissions.
      * @return A {@code Permissions} object created from the given bitmask.
      */
-    static Permissions fromBitmask(int bitmask) {
+    static Permissions fromBitmask(long bitmask) {
         return fromBitmask(bitmask, 0);
     }
 
@@ -104,7 +104,7 @@ public interface Permissions {
      * @param deniedBitmask  The bitmask of denied permissions.
      * @return A {@code Permissions} object created from the given bitmasks.
      */
-    static Permissions fromBitmask(int allowedBitmask, int deniedBitmask) {
+    static Permissions fromBitmask(long allowedBitmask, long deniedBitmask) {
         PermissionsBuilder permissionsBuilder = new PermissionsBuilder();
 
         for (PermissionType permissionType : PermissionType.values()) {
